fix: move migrate helpers to auth package
