+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/utils/migrate-idp-resources.ts
