+++ packages/amplify-provider-awscloudformation/package.json
@@ -28,6 +28,7 @@
     "extract-api": "ts-node ../../scripts/extract-api.ts"
   },
   "dependencies": {
+    "@aws-amplify/amplify-category-auth": "3.2.3",
     "@aws-amplify/amplify-category-custom": "3.0.8",
     "@aws-amplify/amplify-cli-core": "4.0.8",
     "@aws-amplify/amplify-cli-logger": "1.3.2",
