+++ packages/amplify-provider-awscloudformation/src/push-resources.ts
@@ -81,7 +81,7 @@ import { printCdkMigrationWarning } from './print-cdk-migration-warning';
 import { minifyJSONFile } from './utils/minify-json';
 import { handleCloudFormationError } from './cloud-formation-error-handler';
 import { handleCommonSdkError } from './handle-common-sdk-errors';
-import { exportHostedUIProvidersFromCurrCloudRootStack, migrateResourcesToCfn } from './utils/migrate-idp-resources';
+import { exportHostedUIProvidersFromCurrCloudRootStack, migrateResourcesToCfn } from '@aws-amplify/amplify-category-auth';
 
 const logger = fileLogger('push-resources');
